+++ Src/Core/VpnHood.Core.Client/ClientUdpChannelFactory.cs
@@ -23,10 +23,11 @@ public class ClientUdpChannelFactory
         var udpClient = options.SocketFactory.CreateUdpClient(options.RemoteEndPoint.AddressFamily);
         ClientUdpChannelTransmitter? transmitter = null;
         try {
+            var zz = udpClient.Client;
             transmitter = new ClientUdpChannelTransmitter(udpClient, options.ServerKey);
             var udpChannel = new UdpChannel(transmitter, options);
-            transmitter.UdpChannel = udpChannel;
             transmitter.BufferSize = options.BufferSize;
+            transmitter.UdpChannel = udpChannel;
             return udpChannel;
         }
         catch {